backend/requirements.txt
beautifulsoup4==4.6.3
Flask==1.0.2
PyYAML==3.13
redis==2.10.6
requests==2.20.1
rq==0.12.0
